add nmstate operator
Relies on a singleton nmstate instance living in openshift-nmstate in
order to function properly The singleton NMState resource has to exist
before the APIs are available, it seems, which is why it's included in
base with a very broad node selector. The CRs can select which nodes to
apply to via their own node selectors.